added SettingsParser class, modified settings class and TestSuite
